Refactor debugger client
